change the default for thmin (threshold minimum)
But when I set as 2, no cluster was obtained. Then I tried some other numbers. when it is set as 0, then the whole hemisphere will be the cluster. When the number is between 0.001 and 0.35 (roughly), the cluster is the same size as the label. As the activation level is set as larger than 0.35, the size of cluster becomes smaller. Is this activation level similar to the significance (i.e., 1.3 represents 0.05)?